feat: mock show transaction level
